+++ autoarray/plot/visuals/one_d.py
